+++ app/models/early_career_payments/eligibility.rb
@@ -2,9 +2,13 @@ module EarlyCareerPayments
   class Eligibility < ApplicationRecord
     EDITABLE_ATTRIBUTES = [
       :nqt_in_academic_year_after_itt,
-      :employed_as_supply_teacher
+      :employed_as_supply_teacher,
+      :has_entire_term_contract
     ].freeze
     AMENDABLE_ATTRIBUTES = [].freeze
+    ATTRIBUTE_DEPENDENCIES = {
+      "employed_as_supply_teacher" => ["has_entire_term_contract"]
+    }.freeze
 
     self.table_name = "early_career_payments_eligibilities"
 
@@ -12,6 +16,7 @@ module EarlyCareerPayments
 
     validates :nqt_in_academic_year_after_itt, on: [:"nqt-in-academic-year-after-itt", :submit], inclusion: {in: [true, false], message: "Select yes if you did your NQT in the academic year after your ITT"}
     validates :employed_as_supply_teacher, on: [:"supply-teacher", :submit], inclusion: {in: [true, false], message: "Select yes if you are currently employed as a supply teacher"}
+    validates :has_entire_term_contract, on: [:"entire-term-contract", :submit], inclusion: {in: [true, false], message: "Select yes if you have a contract to teach at the same school for one term or longer"}, if: :employed_as_supply_teacher?
 
     def policy
       EarlyCareerPayments
@@ -19,11 +24,19 @@ module EarlyCareerPayments
 
     def ineligible?
       ineligible_nqt_in_academic_year_after_itt? ||
-        not_employed_directly?
+        no_entire_term_contract?
     end
 
     def award_amount
       BigDecimal("2000.00")
+    end
+
+    def reset_dependent_answers
+      ATTRIBUTE_DEPENDENCIES.each do |attribute_name, dependent_attribute_names|
+        dependent_attribute_names.each do |dependent_attribute_name|
+          write_attribute(dependent_attribute_name, nil) if changed.include?(attribute_name)
+        end
+      end
     end
 
     private
@@ -32,8 +45,8 @@ module EarlyCareerPayments
       nqt_in_academic_year_after_itt == false
     end
 
-    def not_employed_directly?
-      employed_as_supply_teacher == true
+    def no_entire_term_contract?
+      employed_as_supply_teacher? && has_entire_term_contract == false
     end
   end
 end