+++ app/models/early_career_payments/slug_sequence.rb
@@ -12,6 +12,7 @@ module EarlyCareerPayments
     SLUGS = [
       "nqt-in-academic-year-after-itt",
       "supply-teacher",
+      "entire-term-contract",
       "check-your-answers",
       "ineligible"
     ].freeze
@@ -24,6 +25,7 @@ module EarlyCareerPayments
 
     def slugs
       SLUGS.dup.tap do |sequence|
+        sequence.delete("entire-term-contract") unless claim.eligibility.employed_as_supply_teacher?
         sequence.delete("ineligible") unless claim.eligibility.ineligible?
       end
     end